backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
pdfplumber==0.11.0
python-docx==1.1.0
python-multipart==0.0.9
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.0
stripe==10.0.0
pydantic==2.9.0
python-jose==3.3.0
httpx==0.27.0
